+++ src/main/java/uk/q3c/krail/core/ui/form/BeanFieldGroupBase.java
@@ -20,11 +20,11 @@ import com.vaadin.data.fieldgroup.FieldGroup;
 import com.vaadin.data.util.BeanItem;
 import com.vaadin.ui.Field;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import uk.q3c.krail.core.data.KrailEntity;
 import uk.q3c.krail.core.i18n.I18NProcessor;
 import uk.q3c.krail.core.option.Option;
 import uk.q3c.krail.core.option.OptionContext;
 import uk.q3c.krail.core.validation.BeanValidator;
+import uk.q3c.krail.persist.KrailEntity;
 
 import java.util.HashMap;
 import java.util.Map;