+++ src/main/java/uk/q3c/krail/i18n/I18NKeyConverter.java
@@ -11,7 +11,7 @@
  *
  */
 
-package uk.q3c.krail.core.data;
+package uk.q3c.krail.i18n;
 
 import uk.q3c.krail.core.i18n.I18NKey;
 import uk.q3c.util.data.ConversionException;