+++ src/test/groovy/uk/q3c/krail/config/bind/ApplicationConfigurationModuleTest.groovy
@@ -17,16 +17,17 @@ import com.google.common.collect.Lists
 import com.google.inject.*
 import spock.lang.Specification
 import uk.q3c.krail.config.config.IniFileConfig
-import uk.q3c.krail.core.data.DataModule
 import uk.q3c.krail.core.guice.uiscope.UIScopeModule
 import uk.q3c.krail.core.persist.inmemory.common.InMemoryModule
 import uk.q3c.krail.core.services.ServicesModule
+import uk.q3c.krail.core.vaadin.DataModule
 import uk.q3c.krail.testutil.eventbus.TestEventBusModule
 import uk.q3c.krail.testutil.guice.vsscope.TestVaadinSessionScopeModule
 import uk.q3c.krail.testutil.i18n.TestI18NModule
 import uk.q3c.krail.testutil.option.TestOptionModule
 import uk.q3c.krail.util.UtilsModule
 import uk.q3c.util.UtilModule
+
 /**
  * Created by David Sowerby on 15 Jan 2016
  */