+++ src/test/groovy/uk/q3c/krail/persist/jpa/SelectTest.java
@@ -11,15 +11,16 @@
  *
  */
 
-package uk.q3c.krail.core.data;
+package uk.q3c.krail.persist.jpa;
 
 import com.mycila.testing.junit.MycilaJunitRunner;
 import com.mycila.testing.plugin.guice.GuiceContext;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import uk.q3c.krail.core.data.TestEntity;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static uk.q3c.krail.core.data.Select.Compare.GREATER_THAN;
+import static org.assertj.core.api.Assertions.*;
+import static uk.q3c.krail.persist.jpa.Select.Compare.*;
 
 @RunWith(MycilaJunitRunner.class)
 @GuiceContext({})